requirements.txt
aiofiles==0.8.0
anyio==3.6.1
arrow==1.2.2
asgiref==3.5.2
binaryornot==0.4.4
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cookiecutter==1.7.3
docopt==0.6.2
fastapi==0.78.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
Jinja2==3.1.2
jinja2-time==0.2.0
loguru==0.6.0
Markdown==3.3.7
MarkupSafe==2.1.1
msgpack==1.0.4
multidict==6.0.2
nb-cli==0.6.7
nonebot-adapter-onebot==2.1.1
nonebot2==2.0.0b4
Pillow==9.1.1
pipreqs==0.4.11
playwright==1.22.0
poyo==0.5.0
prompt-toolkit==3.0.30
pydantic==1.9.1
pyee==8.1.0
pyfiglet==0.8.post1
pygtrie==2.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
text-unidecode==1.3
tomlkit==0.10.2
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.1
yarg==0.1.9
yarl==1.7.2
